py: Rename listener -> endpoint
