+++ playground/control/api/__init__.py
@@ -1,6 +1,7 @@
+
 # -*- coding: utf-8 -*-
 
-from .listener import PlaygroundAPI
+from .endpoint import PlaygroundAPI
 from .handler import PlaygroundEventHandler
